Moved long function away from server.ts to declutter
